requirements.txt
# Server and WebSockets
Flask==2.2.3
Flask-SocketIO==5.3.4
eventlet==0.33.3
python-engineio==4.4.1
python-socketio==5.8.0

# Speech-to-Text
openai-whisper==20231117
torch==2.0.1
torchaudio==2.0.2
numpy==1.24.3

# Text-to-Speech
gTTS==2.3.2

# LLM dependencies
transformers==4.30.2
accelerate==0.20.3
bitsandbytes==0.40.2
safetensors==0.3.1
scipy==1.10.1
peft==0.4.0

# Utilities
python-dateutil==2.8.2
pydub==0.25.1
requests==2.31.0